Handle not found man pages
By default when you call 'man in Man-mode it opens a new buffer, then
starts a background process to populate the buffer with the
manpage. Later (although usually not *much* later) when that
background process exits, if there was no information found then
Man-mode will kill the buffer it created and send a 'message
indicating that nothing was found. That makes sense when we run it
interactively but when run from the command line the result is that
there's just no buffer and a quickly disappearing message telling you
why.

So we add a hook to 'kill-buffer-hook that detects when we're killing
an empty buffer in Man-mode and then if we are it creates and displays
a new buffer also in Man-mode with an approproate message as to why.